Updates profile framework and database setup
Updates the profile framework by removing the foreign key constraint on title_id in the bookmark and rating_history tables. This allows for titles to be bookmarked and rated even if they are not present in the main movie database.

Changes the default dotenv filename to ".env" and adds a new recipe "neondb-create" to the justfile for setting up a Neon database.